[Encoding][NFC] Move non attribute implementation to EncodingTypes.cpp
It follows the convention, which only put the attribute implementation
and local functions to EncodingAttr.cpp. The implementation of other
utilities are moved to EncodingTypes.cpp.
