+++ compiler/src/iree/compiler/Dialect/Encoding/IR/BUILD.bazel
@@ -50,6 +50,7 @@ iree_compiler_cc_library(
         "EncodingOps.cpp",
         "EncodingOps.cpp.inc",
         "EncodingTypeInterfaces.cpp.inc",
+        "EncodingTypes.cpp",
         "EncodingTypes.cpp.inc",
     ],
     hdrs = [
