+++ compiler/src/iree/compiler/Dialect/Encoding/IR/CMakeLists.txt
@@ -34,6 +34,7 @@ iree_cc_library(
     "EncodingOps.cpp"
     "EncodingOps.cpp.inc"
     "EncodingTypeInterfaces.cpp.inc"
+    "EncodingTypes.cpp"
     "EncodingTypes.cpp.inc"
   DEPS
     ::EncodingEnumsGen
